virtcontainers/cli: refactor code
Fixes #302
