+++ cli/kata-check_amd64.go
@@ -5,22 +5,8 @@
 
 package main
 
-/*
-#include <linux/kvm.h>
-
-const int ioctl_KVM_CREATE_VM = KVM_CREATE_VM;
-*/
-import "C"
-
 import (
-	"syscall"
-
 	"github.com/sirupsen/logrus"
-)
-
-// variables rather than consts to allow tests to modify them
-var (
-	kvmDevice = "/dev/kvm"
 )
 
 // archRequiredCPUFlags maps a CPU flag value to search for and a
@@ -66,58 +52,19 @@ var archRequiredKernelModules = map[string]kernelModule{
 // kvmIsUsable determines if it will be possible to create a full virtual machine
 // by creating a minimal VM and then deleting it.
 func kvmIsUsable() error {
-	flags := syscall.O_RDWR | syscall.O_CLOEXEC
-
-	f, err := syscall.Open(kvmDevice, flags, 0)
-	if err != nil {
-		return err
-	}
-	defer syscall.Close(f)
-
-	fieldLogger := kataLog.WithField("check-type", "full")
-
-	fieldLogger.WithField("device", kvmDevice).Info("device available")
-
-	vm, _, errno := syscall.Syscall(syscall.SYS_IOCTL,
-		uintptr(f),
-		uintptr(C.ioctl_KVM_CREATE_VM),
-		0)
-	if errno != 0 {
-		if errno == syscall.EBUSY {
-			fieldLogger.WithField("reason", "another hypervisor running").Error("cannot create VM")
-		}
-
-		return errno
-	}
-	defer syscall.Close(int(vm))
-
-	fieldLogger.WithField("feature", "create-vm").Info("feature available")
-
-	return nil
+	return genericKvmIsUsable()
 }
 
 func archHostCanCreateVMContainer() error {
 	return kvmIsUsable()
 }
 
+// hostIsVMContainerCapable checks to see if the host is theoretically capable
+// of creating a VM container.
+func hostIsVMContainerCapable(details vmContainerCapableDetails) error {
+	return genericHostIsVMContainerCapable(details)
+}
+
 func archKernelParamHandler(onVMM bool, fields logrus.Fields, msg string) bool {
-	param, ok := fields["parameter"].(string)
-	if !ok {
-		return false
-	}
-
-	// This option is not required when
-	// already running under a hypervisor.
-	if param == "unrestricted_guest" && onVMM {
-		kataLog.WithFields(fields).Warn(kernelPropertyCorrect)
-		return true
-	}
-
-	if param == "nested" {
-		kataLog.WithFields(fields).Warn(msg)
-		return true
-	}
-
-	// don't ignore the error
-	return false
+	return genericArchKernelParamHandler(onVMM, fields, msg)
 }
